requirements.txt
# LangChain Core
langchain
langchain-core~=0.3.63
langchain-community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn~=1.6.1

#gradio
gradio~=5.31.0

# Data Visualization
matplotlib~=3.10.3
